More AI States
Implemented AI Attack and Flee states, learned about regions to make the AI Controller script more readable, and the Patrol state is ready but adding waypoints is being problematic. Also created a prefab for the AI Tank and got it to spawn in like the player through the game manager. AI Tank has a ton of variables for designers to mess with.