AI Tweaks and Powerups
Gave Sharp Ears some room to work with so it's not jittery while locked onto the player, added three fully functioning pickups/powerups, one restores the tank's health, one increases the tank's move and turn speed, and one makes the tank's next shot stun opposing tanks. Changed TankPawn to make it stunnable, made the prefabs for the pickups, pickup spawners, and stun bullet (with a script that inherits from DamageOnHit), eliminated damage variable in TankPawn so the data is stored in the bullets, and revamped the PowerupManager so that the player can't pick up health pickups if they're at full hp, can't pick up stun pickups if the powerup already exists in the powerups list, and made sure picking up additional speed pickups only resets the timer for the one in the powerup list rather than adding a new powerup to the list.