UI Upgrade
Gave the player much better visuals, including a health bar that expands when max health increases and an icon that tells them when their cannon is on cooldown as well as if they've picked up the stun projectile. Also broke TankPawn into PlayerTankPawn & AITankPawn and broke Health into PlayerHealth & AIHealth to make better use of inheritance and remove fields that were irrelevant to a given pawn.