Changed to new computer
Some things might appear changed simply because of the switch to the new computer, but now the player will be able to play again if they go back to the menu and hit play. This will also be my reset point if adding audio sources to my objects instead of using the Audio Manager doesn't work either.